refactor: move FeatureProvider to components
