+++ packages/bezier-react/src/components/FeatureProvider/FeatureProvider.types.ts
@@ -1,0 +1,18 @@
+import {
+  type Feature,
+  type FeatureType,
+} from '~/src/features'
+import { type ChildrenProps } from '~/src/types/props'
+
+export type FeatureFlag = Record<FeatureType, boolean>
+
+interface FeatureProviderOwnProps {
+  /**
+   * Features to activate.
+   */
+  features: Feature[]
+}
+
+export interface FeatureProviderProps extends
+  ChildrenProps,
+  FeatureProviderOwnProps {}
