+++ packages/bezier-react/src/components/FeatureProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './FeatureProvider'
+export * from './FeatureProvider.types'
