+++ packages/bezier-react/src/features/index.ts
@@ -1,10 +1,2 @@
-export type { Feature } from './Feature'
-export { FeatureType } from './Feature'
-
-export {
-  FeatureProvider,
-  useFeatureFlag,
-  type FeatureFlag,
-} from './FeatureProvider'
-
-export { SmoothCornersFeature } from './SmoothCornersFeature'
+export * from './Feature'
+export * from './SmoothCornersFeature'
